Move sanitation and validation into one package
