+++ src/main/java/org/parchmentmc/compass/data/validation/Validator.java
@@ -1,4 +1,4 @@
-package org.parchmentmc.compass.validation;
+package org.parchmentmc.compass.data.validation;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.gradle.api.Named;
