+++ src/main/java/org/parchmentmc/compass/data/validation/impl/BridgeValidator.java
@@ -1,13 +1,12 @@
-package org.parchmentmc.compass.validation.impl;
+package org.parchmentmc.compass.data.validation.impl;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
-import org.parchmentmc.compass.validation.AbstractValidator;
-import org.parchmentmc.compass.validation.ValidationIssue;
+import org.parchmentmc.compass.data.validation.AbstractValidator;
+import org.parchmentmc.compass.data.validation.ValidationIssue;
 import org.parchmentmc.feather.metadata.ClassMetadata;
 import org.parchmentmc.feather.metadata.MethodMetadata;
 import org.parchmentmc.feather.util.AccessFlag;
 
-import java.util.List;
 import java.util.function.Consumer;
 
 import static org.parchmentmc.feather.mapping.MappingDataContainer.*;
