+++ src/main/java/org/parchmentmc/compass/data/validation/impl/ParameterIndexValidator.java
@@ -1,9 +1,9 @@
-package org.parchmentmc.compass.validation.impl;
+package org.parchmentmc.compass.data.validation.impl;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.parchmentmc.compass.util.DescriptorIndexer;
-import org.parchmentmc.compass.validation.AbstractValidator;
-import org.parchmentmc.compass.validation.ValidationIssue;
+import org.parchmentmc.compass.data.validation.AbstractValidator;
+import org.parchmentmc.compass.data.validation.ValidationIssue;
 import org.parchmentmc.feather.mapping.MappingDataContainer.ClassData;
 import org.parchmentmc.feather.mapping.MappingDataContainer.MethodData;
 import org.parchmentmc.feather.mapping.MappingDataContainer.ParameterData;
