+++ src/main/java/org/parchmentmc/compass/tasks/SanitizeData.java
@@ -10,12 +10,12 @@ import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.InputDirectory;
 import org.gradle.api.tasks.TaskAction;
 import org.parchmentmc.compass.CompassPlugin;
-import org.parchmentmc.compass.sanitation.DataSanitizer;
-import org.parchmentmc.compass.sanitation.Sanitizer;
-import org.parchmentmc.compass.sanitation.impl.BouncerDataMover;
-import org.parchmentmc.compass.sanitation.impl.DescriptorParametersSanitizer;
-import org.parchmentmc.compass.sanitation.impl.EnumValueOfRemover;
-import org.parchmentmc.compass.sanitation.impl.SyntheticsRemover;
+import org.parchmentmc.compass.data.sanitation.DataSanitizer;
+import org.parchmentmc.compass.data.sanitation.Sanitizer;
+import org.parchmentmc.compass.data.sanitation.impl.BouncerDataMover;
+import org.parchmentmc.compass.data.sanitation.impl.DescriptorParametersSanitizer;
+import org.parchmentmc.compass.data.sanitation.impl.EnumValueOfRemover;
+import org.parchmentmc.compass.data.sanitation.impl.SyntheticsRemover;
 import org.parchmentmc.compass.storage.io.MappingIOFormat;
 import org.parchmentmc.compass.util.download.BlackstoneDownloader;
 import org.parchmentmc.feather.mapping.MappingDataContainer;
