+++ src/main/java/org/parchmentmc/compass/tasks/ValidateData.java
@@ -13,9 +13,9 @@ import org.parchmentmc.compass.CompassPlugin;
 import org.parchmentmc.compass.storage.io.MappingIOFormat;
 import org.parchmentmc.compass.util.ResultContainer;
 import org.parchmentmc.compass.util.download.BlackstoneDownloader;
-import org.parchmentmc.compass.validation.ValidationIssue;
-import org.parchmentmc.compass.validation.action.DataValidator;
-import org.parchmentmc.compass.validation.impl.*;
+import org.parchmentmc.compass.data.validation.ValidationIssue;
+import org.parchmentmc.compass.data.validation.action.DataValidator;
+import org.parchmentmc.compass.data.validation.impl.*;
 import org.parchmentmc.feather.mapping.MappingDataContainer;
 import org.parchmentmc.feather.metadata.SourceMetadata;
 import org.slf4j.Marker;
